[mocks] some renaming
