+++ test/page-encointer/meetup/attestation/components/stateMachinePartyA_test.dart
@@ -5,7 +5,7 @@ import 'package:encointer_wallet/service/substrateApi/api.dart';
 import 'package:encointer_wallet/store/app.dart';
 import 'package:encointer_wallet/store/encointer/types/attestationState.dart';
 
-import 'package:encointer_wallet/mocks/api/apiEncointer_mock.dart';
+import 'package:encointer_wallet/mocks/api/apiEncointer.dart';
 import 'package:encointer_wallet/mocks/data/mockAccountData.dart';
 import 'package:encointer_wallet/mocks/data/mockEncointerData.dart';
