+++ test/store/app_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:encointer_wallet/store/app.dart';
 
-import 'package:encointer_wallet/mocks/localStorage_mock.dart';
+import 'package:encointer_wallet/mocks/localStorage.dart';
 
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
